presentation-api/controlling-ua/defaultRequest_error.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Presentation API, testing to start a new presentation with an default request setup. (error)</title>
<link rel="author" title="Marius Wessel" href="http://www.fokus.fraunhofer.de">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-controlling-user-agent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    // -------------------
    // defaultRequest init
    // -------------------
    var validUnixDate = new Date().getTime() + String(Math.floor(Math.random() * 1e5)),
            validPresURL = '../receiving-ua/idlharness.html#__castAppId__=2334D33A/__castClientId__=' + validUnixDate;
            apiPreDefinedErrorMessage = 'PresentationRequest::start() requires user gesture.';

    navigator.presentation.defaultRequest = new PresentationRequest(validPresURL);
    // ------------------------------
    // Start New Presentation with
    // 'default request' Test - BEGIN
    // ------------------------------
    promise_test(function () {
        return navigator.presentation.defaultRequest.start()
                .catch(function (err) {

                    assert_equals(err.message, apiPreDefinedErrorMessage, 'The error message is "PresentationRequest::start() requires user gesture."');

                    assert_equals(err.name, 'InvalidAccessError', 'The error name is InvalidAccessError.');

                });
    }, "The presentation could not start, because a user gesture is required.");
    // ----------------------------
    // Start New Presentation with
    // 'default request' Test - END
    // ----------------------------
</script>
